[DABOM-494] domain 이름 eventoutbox로 변경
* DABOM-491 refactor: domain 이름 eventoutbox로 변경

* DABOM-494 refactor: 남아있는 usage 제거

* DABOM-494 refactor: mysql 주소 원상복구

* DABOM-494 style: checkStyle 적용

* DABOM-494 refactor: test 코드 오류 수정

* DABOM-494 refactor: test 코드 Long 형변환 추가

+++ src/test/java/com/template/worker/jobs/reconciliation/step/InvalidateFamilyInfoAndRemainingStepConfigTest.java
@@ -90,7 +90,7 @@ class InvalidateFamilyInfoAndRemainingStepConfigTest {
                         properties,
                         new BatchRetrySupport(3, 0L));
 
-        when(reader.read()).thenReturn(10L, null);
+        when(reader.read()).thenReturn(10L, (Long) null);
         doThrow(new RedisConnectionFailureException("redis")).when(writer).write(any());
 
         Step step = stepConfig.invalidateFamilyInfoAndRemainingStep();
